fix: Resolve module resolution error and apply pending refactors
This commit addresses a critical runtime error that prevented the application from rendering. The error, "The requested module does not provide an export named...", was caused by a violation of the `react-refresh/only-export-components` lint rule in the `LanguageContext.tsx` file.

The fix involves a significant refactoring of the language context:
- The `useLanguage` hook has been moved to its own file in `src/hooks/`.
- The `LanguageContext` object has been moved to its own file.
- All 16 files that used the hook have had their import paths updated.

This change resolves the underlying issue with the development server's module resolution.

Per user instruction, this commit also includes all other work from this session:
- refactor(qo-c003): Update item details page with detailed content.
- refactor(qo-c002): Extract a reusable `MenuItemCard` component.
- fix(qo-c002): Align 'Add to Cart' button to the bottom of the card.
