+++ src/contexts/LanguageContextObject.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import type { LanguageContextType } from './LanguageContext';
+
+export const LanguageContext = createContext<LanguageContextType | undefined>(undefined);
