+++ src/pages/qo_c007_order_status.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import { Clock, CheckCircle, Phone, MessageSquare, Bell, BellOff } from 'lucide-react';
 import PageLayout from '../components/PageLayout';
-import { useLanguage } from '../contexts/LanguageContext';
+import { useLanguage } from '../hooks/useLanguage';
 
 const QOOrderStatus = () => {
   const { language } = useLanguage();
