+++ src/pages/qo_s001_staff_login.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { User, Lock } from 'lucide-react';
 import PageLayout from '../components/PageLayout';
-import { useLanguage } from '../contexts/LanguageContext';
+import { useLanguage } from '../hooks/useLanguage';
 
 const QOStaffLogin = () => {
   const { language } = useLanguage();
