+++ src/pages/qo_s002_dashboard.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import { Bell, Settings, LogOut } from 'lucide-react';
 import PageLayout from '../components/PageLayout';
-import { useLanguage } from '../contexts/LanguageContext';
+import { useLanguage } from '../hooks/useLanguage';
 
 const QOStaffDashboard = () => {
   const { language } = useLanguage();
