+++ src/pages/qo_s003_order_management.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { Search, RefreshCw, Bell } from 'lucide-react';
 import PageLayout from '../components/PageLayout';
-import { useLanguage } from '../contexts/LanguageContext';
+import { useLanguage } from '../hooks/useLanguage';
 
 const QOOrderManagement = () => {
   const { language } = useLanguage();
