+++ src/pages/qo_s004_kitchen_display.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { useLanguage } from '../contexts/LanguageContext';
+import { useLanguage } from '../hooks/useLanguage';
 
 interface OrderItem {
   name: string;
